requirements.txt
<<<<<<< Updated upstream:requirements.txt
aiohttp==3.6.2
=======
#aiohttp==3.5.4
>>>>>>> Stashed changes:src/requirements.txt
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.4.4
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
<<<<<<< Updated upstream:requirements.txt
discord==1.0.1
=======
#discord==1.0.1
>>>>>>> Stashed changes:src/requirements.txt
discord.py==1.3.4
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
filetype==1.0.5
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-vision==0.41.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
idna==2.8
multidict==4.7.2
pillow==2.2.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
websockets==6.0
yarl==1.4.2
